refactor: 전역예외처리 패키지 global로 이동
