+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/auth/cookie/ProdCookieUtils.java
@@ -1,0 +1,4 @@
+package kakaotech.bootcamp.respec.specranking.domain.auth.cookie;
+
+public class ProdCookieUtils {
+}
