+++ src/main/java/kakaotech/bootcamp/respec/specranking/global/exception/DuplicateNicknameException.java
@@ -1,4 +1,4 @@
-package kakaotech.bootcamp.respec.specranking.domain.user.util;
+package kakaotech.bootcamp.respec.specranking.global.exception;
 
 public class DuplicateNicknameException extends RuntimeException {
     public DuplicateNicknameException(String message) {
